Moved background task initialization to Actions
